requirements.txt
asgiref==3.4.1
Django==3.2.13
django-phonenumber-field==6.0.0
phonenumbers==8.12.50
Pillow==8.4.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
